+++ misskey-api/src/endpoint/channels/followed.rs
@@ -64,11 +64,11 @@ mod tests {
     async fn request_paginate() {
         let client = TestClient::new();
         let channel = client
-            .test(crate::endpoint::channels::create::Request {
-                name: "test channel".to_string(),
-                description: None,
-                banner_id: None,
-            })
+            .test(
+                crate::endpoint::channels::create::Request::builder()
+                    .name("test channel")
+                    .build(),
+            )
             .await;
         client
             .test(crate::endpoint::channels::follow::Request {