+++ misskey-api/src/endpoint/drive/files/attached_notes.rs
@@ -23,23 +23,11 @@ mod tests {
         let client = TestClient::new();
         let file = client.create_text_file("test.txt", "test").await;
         client
-            .test(crate::endpoint::notes::create::Request {
-                visibility: None,
-                visible_user_ids: None,
-                text: None,
-                cw: None,
-                via_mobile: None,
-                local_only: None,
-                no_extract_mentions: None,
-                no_extract_hashtags: None,
-                no_extract_emojis: None,
-                file_ids: Some(vec![file.id.clone()]),
-                reply_id: None,
-                renote_id: None,
-                poll: None,
-                #[cfg(feature = "12-47-0")]
-                channel_id: None,
-            })
+            .test(
+                crate::endpoint::notes::create::Request::builder()
+                    .file_ids(vec![file.id.clone()])
+                    .build(),
+            )
             .await;
 
         client.test(Request { file_id: file.id }).await;