+++ misskey-api/src/endpoint/i.rs
@@ -18,6 +18,10 @@ pub mod user_group_invites;
 #[cfg_attr(docsrs, doc(cfg(feature = "12-67-0")))]
 pub mod registry;
 
+#[cfg(feature = "12-79-0")]
+#[cfg_attr(docsrs, doc(cfg(feature = "12-79-0")))]
+pub mod gallery;
+
 #[derive(Serialize, Default, Debug, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct Request {}