+++ misskey-api/src/endpoint/i/update.rs
@@ -52,6 +52,11 @@ pub struct Request {
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub is_explorable: Option<bool>,
+    #[cfg(feature = "12-77-0")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "12-77-0")))]
+    #[serde(skip_serializing_if = "Option::is_none")]
+    #[builder(default, setter(strip_option))]
+    pub hide_online_status: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(strip_option))]
     pub careful_bot: Option<bool>,
@@ -149,6 +154,8 @@ mod tests {
                 is_locked: Some(true),
                 #[cfg(feature = "12-63-0")]
                 is_explorable: Some(false),
+                #[cfg(feature = "12-77-0")]
+                hide_online_status: Some(true),
                 careful_bot: Some(true),
                 auto_accept_followed: Some(true),
                 is_bot: Some(true),
@@ -201,6 +208,8 @@ mod tests {
                 is_locked: None,
                 #[cfg(feature = "12-63-0")]
                 is_explorable: None,
+                #[cfg(feature = "12-77-0")]
+                hide_online_status: None,
                 careful_bot: None,
                 auto_accept_followed: None,
                 is_bot: None,